requirements.txt
numpy==1.25.1
Pillow==10.0.0
scikit_learn==1.3.0
seaborn==0.12.2
segmentation_models_pytorch==0.3.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
tensorboard
opencv-python
